+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [0.2.1] - 2025-08-15
+
+### Fixed
+- Improved many design elements to be more consitant and user friendly
+
+### Changed
+- Improved many design elements to be more consistant and user friendly.
+
 ## [0.2.0] - 2025-08-15
 
 ### Changed